+++ front/lib/actions/mcp.ts
@@ -11,7 +11,7 @@ import { tryCallMCPTool } from "@app/lib/actions/mcp_actions";
 import {
   augmentInputsWithConfiguration,
   hideInternalConfiguration,
-} from "@app/lib/actions/mcp_internal_actions/input_schemas";
+} from "@app/lib/actions/mcp_internal_actions/input_schema_parsing";
 import { getMCPEvents } from "@app/lib/actions/pubsub";
 import type { DataSourceConfiguration } from "@app/lib/actions/retrieval";
 import type { TableDataSourceConfiguration } from "@app/lib/actions/tables_query";
