+++ front/lib/actions/mcp_internal_actions/input_schema_parsing.ts
